requirements.txt
jsonschema
pyyaml
pytest
Django>=4.2
djangorestframework
drf-yasg
uvicorn